App menu

+++ src/main/java/link/locutus/discord/commands/manager/v2/impl/pw/commands/AppMenuCommands.java
@@ -5,6 +5,7 @@ import link.locutus.discord.commands.manager.v2.command.*;
 import link.locutus.discord.commands.manager.v2.impl.discord.DiscordChannelIO;
 import link.locutus.discord.commands.manager.v2.impl.discord.DiscordHookIO;
 import link.locutus.discord.commands.manager.v2.impl.discord.permission.RolePermission;
+import link.locutus.discord.commands.manager.v2.impl.pw.refs.CM;
 import link.locutus.discord.db.GuildDB;
 import link.locutus.discord.db.entities.menu.AppMenu;
 import link.locutus.discord.db.entities.menu.MenuState;
@@ -19,7 +20,42 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.TimeUnit;
 
 public class AppMenuCommands {
-    public static final PassiveExpiringMap<Long, AppMenu> USER_MENU_STATE = new PassiveExpiringMap<Long, AppMenu>(2, TimeUnit.MINUTES);
+    public static final PassiveExpiringMap<Long, AppMenu> USER_MENU_STATE = new PassiveExpiringMap<Long, AppMenu>(15, TimeUnit.MINUTES);
+
+//    public boolean onButtonUse(IMessageIO io, GuildDB db, User user, String buttonLabel) {
+//        AppMenu menu = getAppMenu(user, io, false);
+//        if (menu == null) {
+//            return false;
+//        }
+//        MenuState state = menu.state;
+//        switch (state) {
+//            case REORDER:
+//                if (menu.lastPressedButton != null) {
+//                    swapMenuButtons(io, db, user, menu, menu.lastPressedButton, buttonLabel);
+//                    break;
+//                }
+//                menu.lastPressedButton = buttonLabel.toLowerCase(Locale.ROOT);
+//                String msg = """
+//                        You are now in reordering mode. Click the buttons to swap their positions. Click the same button again to exit reordering mode.
+//                        Press `cancel` to exit.
+//                        """;
+//                menu.setState(MenuState.REORDER).message().buttons(true).cancel(true).description(msg).queue(io);
+//                break;
+//            case REMOVE_BUTTON:
+//                removeMenuButton(io, db, user, menu, buttonLabel);
+//                break;
+//            case RENAME_BUTTON:
+//                // Rename the button
+//                menu.lastPressedButton = buttonLabel; // Store the button to be renamed
+//                // Open a modal to get the new name
+//                CommandRef renameModal = CM.menu.button.rename.cmd.menu(menu.title).label(buttonLabel).new_label("");
+//                io.create().modal(CommandBehavior.EPHEMERAL, renameModal, "Rename Modal").send();
+//                break;
+//            default:
+//                return false;
+//        }
+//        return true;
+//    }
 
     public AppMenu getAppMenu(User user, IMessageIO io, boolean deleteIfDifferentChannel) {
         long channelId = 0;
@@ -60,60 +96,12 @@ public class AppMenuCommands {
         return false;
     }
 
-    // Placeholder for now, don't code these now
-    public static ICommand OPEN_MENU = null;
-    public static ICommand EDIT_MENU = null;
-    public static ICommand ADD_BUTTON = null;
-    public static ICommand REMOVE_BUTTON = null;
-    public static ICommand RENAME_BUTTON = null;
-    public static ICommand SWAP_BUTTONS = null;
-    public static ICommand NEW_MENU = null;
-    public static ICommand DELETE_MENU = null;
-    public static ICommand RENAME_MENU = null;
-    public static ICommand DESCRIBE_MENU = null;
-    public static ICommand SET_MENU_STATE = null;
-
-    public boolean onButtonUse(IMessageIO io, GuildDB db, User user, String buttonLabel) {
-        AppMenu menu = getAppMenu(user, io, false);
-        if (menu == null) {
-            return false;
-        }
-        MenuState state = menu.state;
-        switch (state) {
-            case REORDER:
-                if (menu.lastPressedButton != null) {
-                    swapMenuButtons(io, db, user, menu, menu.lastPressedButton, buttonLabel);
-                    break;
-                }
-                menu.lastPressedButton = buttonLabel.toLowerCase(Locale.ROOT);
-                String msg = """
-                        You are now in reordering mode. Click the buttons to swap their positions. Click the same button again to exit reordering mode.
-                        Press `cancel` to exit.
-                        """;
-                menu.setState(MenuState.REORDER).message().buttons(true).cancel(true).description(msg).queue(io);
-                break;
-            case REMOVE_BUTTON:
-                removeMenuButton(io, db, user, menu, buttonLabel);
-                break;
-            case RENAME_BUTTON:
-                // Rename the button
-                menu.lastPressedButton = buttonLabel; // Store the button to be renamed
-                // Open a modal to get the new name
-//                CommandRef renameModal = RENAME_BUTTON.menu(menu.title).label(buttonLabel).new_label("");  // TODO FIXME APP MENU COMMAND
-//                io.create().modal(CommandBehavior.EPHEMERAL, renameModal, "Rename Modal").send();
-                break;
-            default:
-                return false;
-        }
-        return true;
-    }
-
     @Command
     @NoFormat
     @Ephemeral
     public void openMenu(@Me IMessageIO io, @Me GuildDB db, @Me User user, AppMenu menu) {
         menu.lastUsedChannel = io.getIdLong();
-        menu.state = MenuState.NONE;
+        menu.setState(MenuState.NONE);
         USER_MENU_STATE.put(user.getIdLong(), menu);
         menu.message().edit(menu.canEdit(db, user)).buttons(true).queue(io);
     }
@@ -129,6 +117,9 @@ public class AppMenuCommands {
             io.create().confirmation(title, body, command).send();
             return null;
         }
+        if (menu.isDefault()) {
+            return "Cannot delete the default menus. Consider editing them instead.";
+        }
         db.getMenuManager().deleteMenu(menu);
         return "Menu deleted successfully.";
     }
@@ -138,6 +129,9 @@ public class AppMenuCommands {
     @RolePermission(Roles.ADMIN)
     @Ephemeral
     public void renameMenu(@Me IMessageIO io, @Me GuildDB db, @Me User user, AppMenu menu, String name) {
+        if (menu.isDefault()) {
+            throw new IllegalArgumentException("Cannot rename the default menus. Consider creating a new one with " + CM.menu.create.cmd.toSlashMention());
+        }
         menu.setState(MenuState.NONE);
         db.getMenuManager().renameMenu(menu, name);
         menu.message().edit(menu.canEdit(db, user)).buttons(true).queue(io);
@@ -156,9 +150,11 @@ public class AppMenuCommands {
 
     @Command
     @NoFormat
-    @RolePermission(Roles.ADMIN)
     @Ephemeral
     public void setMenuState(@Me IMessageIO io, @Me GuildDB db, @Me User user, AppMenu menu, MenuState state) {
+        if (!Roles.ADMIN.has(user, db.getGuild()) && state != MenuState.NONE) {
+            throw new IllegalArgumentException("You do not have permission to set the menu state");
+        }
         menu.messageState(io, db, user, state);
     }
 
@@ -176,9 +172,8 @@ public class AppMenuCommands {
     @NoFormat
     @RolePermission(Roles.ADMIN)
     @Ephemeral
-    public void addMenuButton(@Me IMessageIO io, @Me JSONObject cmdJson, @Me GuildDB db, @Me User user, AppMenu menu, @MenuLabel String label, @Default String command, @Switch("f") boolean force) {
+    public static void addMenuButton(@Me IMessageIO io, @Me JSONObject cmdJson, @Me GuildDB db, @Me User user, AppMenu menu, @MenuLabel String label, @Default String command, @Switch("f") boolean force) {
         label = label.toLowerCase(Locale.ROOT);
-        menu.setState(MenuState.NONE);
         if (label.equalsIgnoreCase("cancel") || label.equalsIgnoreCase("edit")) {
             menu.setState(MenuState.NONE);
             menu.message().options(true).cancel(true).description("You cannot use `cancel` or `edit` as button labels. Please try again").queue(io);
@@ -198,14 +193,29 @@ public class AppMenuCommands {
             }
             menu.buttons.remove(contains);
             db.getMenuManager().saveMenu(menu);
+        } else if (menu.buttons.size() >= 25) {
+            menu.setState(MenuState.NONE);
+            menu.message().options(true).cancel(true).description("You cannot add more than 25 buttons to a menu. Consider creating a new menu as a sub-menu for some of the buttons").queue(io);
         }
         if (command == null || command.isEmpty()) {
-            String desc = "Enter the SLASH command for the button `" + label + "` as you would normally, within the same channel\nNote: Legacy commands are not supported";
-            menu.message().buttons(true).edit(true).description(desc).queue(io);
+            menu.setState(MenuState.ADD_BUTTON);
+            menu.lastPressedButton = label;
+            String desc = "Enter the SLASH command for the button `" + label + "` as you would normally, within the same channel\n" +
+                    "You can use: `{user}`, `{message}` and `{content}` if the menu is from a user or message interaction" +
+                    "Note: Legacy commands are not supported\n\n" +
+                    "Alternatively, cancel, and run " + CM.menu.button.add.cmd.toSlashMention() + " and specify the command";
+            menu.message().buttons(true).cancel(true).description(desc).queue(io);
         } else {
+            String footer = "";
+            if (menu.isDefault()) {
+                footer = "\n\nYou may need to restart discord to see changes to context menus";
+            }
+
+            menu.setState(MenuState.NONE);
             menu.buttons.put(label, command);
             db.getMenuManager().saveMenu(menu);
-            String desc = "Button `" + label + "` added successfully.\n\nPress a button to remove it.\nPress `cancel` to exit";
+            String desc = "Button `" + label + "` added successfully.\n\n" +
+                    "Press a button to remove it.\nPress `cancel` to exit" + footer;
             menu.message().options(true).cancel(true).description(desc).queue(io);
         }
     }
@@ -218,8 +228,13 @@ public class AppMenuCommands {
         label = label.toLowerCase(Locale.ROOT);
         menu.setState(MenuState.REMOVE_BUTTON);
         if (menu.buttons.remove(label) != null) {
+            String footer = "";
+            if (menu.isDefault()) {
+                footer = "\n\nYou may need to restart discord to see changes to context menus";
+            }
+
             db.getMenuManager().saveMenu(menu);
-            String desc = "Removed button `" + label + "`.\n\nPress a button to remove it.\nPress `cancel` to exit";
+            String desc = "Removed button `" + label + "`.\n\nPress a button to remove it.\nPress `cancel` to exit" + footer;
             menu.message().buttons(true).cancel(true).description(desc).queue(io);
         } else {
             menu.message().buttons(true).cancel(true).description("Error: Button not found. Try again.\nPress a button to remove it.\nPress  `cancel` to exit").queue(io);
@@ -230,38 +245,56 @@ public class AppMenuCommands {
     @NoFormat
     @RolePermission(Roles.ADMIN)
     @Ephemeral
-    public void swapMenuButtons(@Me IMessageIO io, @Me GuildDB db, @Me User user, AppMenu menu, @MenuLabel String label1, @MenuLabel String label2) {
+    public void swapMenuButtons(@Me IMessageIO io, @Me GuildDB db, @Me User user, AppMenu menu, @MenuLabel String label1, @Default @MenuLabel String label2) {
         label1 = label1.toLowerCase(Locale.ROOT);
-        label2 = label2.toLowerCase(Locale.ROOT);
-        if (label1.equals(label2)) {
+        if (label2 != null) label2 = label2.toLowerCase(Locale.ROOT);
+        String initialLabel2 = label2;
+        if (label2 == null) label2 = menu.lastPressedButton;
+
+        if (label2 != null && label1.equals(label2)) {
             menu.clearState().message().options(true).description("Cancelled reordering (pressed button twice)\n\n" +
                     "Select an edit option.\nPress `cancel` to exit").cancel(true).queue(io);
             return;
         }
         String command1 = menu.buttons.get(label1);
-        String command2 = menu.buttons.get(label2);
+        String command2 = label2 == null ? null : menu.buttons.get(label2);
 
-        if (command1 == null || command2 == null) {
+        if (command1 == null || (command2 == null && initialLabel2 == null)) {
             List<String> cannotFind = new ArrayList<>();
             if (command1 == null) {
                 cannotFind.add(label1);
             }
             if (command2 == null) {
+                if (initialLabel2 == null) menu.lastPressedButton = null;
                 cannotFind.add(label2);
             }
             String errorMsg = "Cannot find button(s): " + String.join(", ", cannotFind);
             String desc = errorMsg + "\n\nSelect a button to swap.\nPress `cancel` to exit.";
-            menu.setState(MenuState.REORDER).message().options(true).description(desc).cancel(true).queue(io);
+            menu.state = MenuState.REORDER;
+            menu.message().buttons(true).description(desc).cancel(true).queue(io);
             return;
+        }
+        if (label2 == null) {
+            menu.state = MenuState.REORDER;
+            menu.lastPressedButton = label1;
+            String desc = "Selected `" + label1 + "`. Select a button to swap with.\nPress `cancel` to exit.";
+            menu.message().buttons(true).description(desc).cancel(true).queue(io);
+            return;
+        }
+        String footer = "";
+        if (menu.isDefault()) {
+            footer = "\n\nYou may need to restart discord to see changes to context menus";
         }
 
         menu.buttons.put(label1, command2);
         menu.buttons.put(label2, command1);
 
         db.getMenuManager().saveMenu(menu);
-        menu.setState(MenuState.REORDER).message()
+        menu.setState(MenuState.REORDER);
+        menu.message()
                 .options(true)
-                .description("Successfully swapped `" + label1 + "` and `" + label2 + "`.\n\nSelect a button to swap.\nPress `cancel` to exit.")
+                .description("Successfully swapped `" + label1 + "` and `" + label2 +
+                        "`.\n\nSelect a button to swap.\nPress `cancel` to exit." + footer)
                 .cancel(true).queue(io);
     }
 
@@ -286,10 +319,16 @@ public class AppMenuCommands {
             menu.setState(MenuState.RENAME_BUTTON).message().buttons(true).cancel(true).description(desc).queue(io);
             return;
         }
+        String footer = "";
+        if (menu.isDefault()) {
+            footer = "\n\nYou may need to restart discord to see changes to context menus";
+        }
+
         String cmd1 = menu.buttons.remove(label);
         menu.buttons.put(new_label, cmd1);
         db.getMenuManager().saveMenu(menu);
-        String desc = "Successfully rename `" + label + "` to `" + new_label + "`.\n\nSelect a button to rename.\nPress `cancel` to exit.";
+        String desc = "Successfully rename `" + label + "` to `" + new_label +
+                "`.\n\nSelect a button to rename.\nPress `cancel` to exit." + footer;
         menu.setState(MenuState.RENAME_BUTTON).message().options(true).description(desc).cancel(true).queue(io);
     }
 
@@ -311,7 +350,7 @@ public class AppMenuCommands {
     @Command
     @NoFormat
     @Ephemeral
-    public void cancel(@Me IMessageIO io, @Me GuildDB db, User user, AppMenu menu) {
+    public void cancel(@Me IMessageIO io, @Me GuildDB db, @Me User user, AppMenu menu) {
         if (menu.state == MenuState.NONE) {
             menu.clearState().message().buttons(true).edit(menu.canEdit(db, user)).queue(io);
         } else {
